Updates remaining files from _services rename I missed these in the first pass, since there were a lot.
