+++ tests/test_pymunk_physics_service.py
@@ -5,8 +5,8 @@ import unittest
 
 from pyrite.physics import ColliderComponent, RigidbodyComponent
 from pyrite._physics.shapes import Circle
-from pyrite.services import PhysicsService
-from pyrite.services.physics_service import PymunkPhysicsService
+from pyrite._services import PhysicsService
+from pyrite._services.physics_service import PymunkPhysicsService
 from pyrite.transform import TransformComponent, Transform
